extract distributedrunner from simulation code
